+++ barapost/prober.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "1.18.e"
+__version__ = "1.19.a"
 # Year, month, day
-__last_update_date__ = "2020-05-01"
+__last_update_date__ = "2020-05-06"
 
 # |===== Check python interpreter version =====|