+++ barapost/src/prune_seqs.py
@@ -3,14 +3,15 @@
 
 def prune_seqs(packet, mode, value):
     """
-    Function prunes all sequences in fasta-formatted string, leaving 5'-half of the sequence.
+    Function prunes all sequences in fasta-formatted string.
+    Function prunes sequences from both ends.
 
     :param packet: FASTA data;
     :type packet: str;
     :param mode: available values: 'l' (for length) and 'f' (for fraction).
-        Meaning: if mode is 'l', all sequences are pruned from the 5'-end to position 'value' (value > 1),
-                 if mode is 'f', all sequences are pruned from the 5'-end to position L*value (0<value<1),
-                 where L denotes to length of a sequence.
+        Meaning: if mode is 'l', length of all resulting sequences will be equal to 'value' (value > 1),
+                 if mode is 'f', length of all resulting sequences will be equal to 'L*value' (0<value<1),
+                 where L is length of a sequence.
     :type mode: str;
     :param value: see description if 'mode' parameter above;
     :type value: int or float;
@@ -26,11 +27,15 @@ def prune_seqs(packet, mode, value):
         # end if
 
         def prune(seq, value):
-            try:
-                return seq[: value]
-            except IndexError:
-                return seq
-            # end try
+
+            len_seq = len(seq)
+            if len_seq > value:
+                flank = len_seq - value
+                amend = flank % 2
+                flank //= 2
+                seq = seq[flank + amend : (len_seq - flank)]
+            # end if
+            return seq
         # end def prune
     elif mode == 'f':
         if value < 0.0 + 1e-9 or value > 1.0 - 1e-9:
@@ -38,7 +43,12 @@ def prune_seqs(packet, mode, value):
         # end if
 
         def prune(seq, value):
-            return seq[: int(len(seq) * value)]
+
+            len_seq = len(seq)
+            flank = int(len_seq * value)
+            amend = flank % 2
+            flank = (len_seq - flank) // 2
+            return seq[flank + amend : (len_seq - flank)]
         # end def prune
     else:
         raise ValueError("Invalid 'mode' argument passed to function prune_seqs: '{}'".format(mode))
@@ -53,8 +63,7 @@ def prune_seqs(packet, mode, value):
         is_seq = lambda i: i > id_curr and i < id_next
         # Get all sequence lines of current record
         seq_lines = map(lambda j: lines[j], filter(is_seq, range(len(lines))))
-        seq = "".join(seq_lines) # form a single string
-        seq = prune(seq, value) # prune it
+        seq = prune("".join(seq_lines), value) # form a single string and prune it
         packet += lines[id_curr] + '\n' + seq + '\n' # apend pruned record to packet
     # end for